Standardised storage tests
