+++ tests/integration/storage_dynamodb/test_storage_dynamodb_read.py
@@ -3,7 +3,7 @@ from ..storage_read_generic import StorageReadTestMixin
 from cloudwanderer.storage_connectors.dynamodb import DynamoDbConnector, IndexNotAvailableException
 
 
-class TestStorageMemoryRead(StorageReadTestMixin, unittest.TestCase):
+class TestStorageDynamoDbRead(StorageReadTestMixin, unittest.TestCase):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
